requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==4.2.17
django-bootstrap-v5==1.0.11
django-braces==1.16.0
django-crispy-forms==2.3
django-phonenumber-field==8.0.0
phonenumbers==8.13.52
pillow==11.0.0
soupsieve==2.6
sqlparse==0.5.3
